+++ apps/server/src/shared/infra/console/console-writer/console-writer.module.spec.ts
@@ -1,0 +1,22 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ConsoleWriterService } from './console-writer.service';
+
+describe('ConsoleWriterModule', () => {
+	let module: TestingModule;
+	let service: ConsoleWriterService;
+
+	beforeAll(async () => {
+		module = await Test.createTestingModule({
+			imports: [ConsoleWriterService],
+		}).compile();
+		service = module.get<ConsoleWriterService>(ConsoleWriterService);
+	});
+
+	afterAll(async () => {
+		await module.close();
+	});
+
+	it('should have defined service', () => {
+		expect(service).toBeDefined();
+	});
+});