+++ apps/server/src/shared/infra/file-system/index.ts
@@ -1,0 +1,2 @@
+export * from './file-system.adapter';
+export * from './file-system.module';